WIP : migrate from keycloak to OIDC
